Modification du contructeur de Transactions

+++ src/main/java/project/dao/TransactionDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 import project.model.Transaction;
 
 public interface TransactionDao extends GenericDao<Transaction, Integer> {
-	public List<Transaction> getStatement(Integer accountNumber,LocalDate dateFrom,LocalDate dateTo);
+	public List<Transaction> getStatement(Long accountNumber,LocalDate dateFrom,LocalDate dateTo);
 }
